feat: server logs
