+++ apps/web/src/pages/api/log.ts
@@ -1,8 +1,6 @@
-import { serverLogger } from '../../../datadog'
-import tracer from '../../utils/tracer'
+import { serverLogger } from '../../../winston'
 
 export default async function handler(req, res) {
-  const span = tracer.startSpan('api.hello')
   try {
     serverLogger.info(`[${req.query.status}][${req.query.datadogData}]`, {
       metadata: '',
@@ -22,7 +20,6 @@ export default async function handler(req, res) {
       message: req.query,
       status: 500,
     }
-    span.finish()
 
     return res.status(404).json(errorResponse)
   }
