+++ apps/web/src/utils/log.ts
@@ -1,6 +1,5 @@
 import { Currency } from '@pancakeswap/swap-sdk-core'
-
-import { logger } from '../../datadog'
+import { logger } from 'utils/datadog'
 
 export const logTx = ({ account, hash, chainId }: { account: string; hash: string; chainId: number }) => {
   fetch(`/api/_log/${account}/${chainId}/${hash}`)
