+++ apps/web/tsconfig.json
@@ -14,5 +14,5 @@
     }
   },
   "exclude": ["**/node_modules", "**/.*/", "**/config/abi/types/", "**/abis/types/*.ts"],
-  "include": ["next-env.d.ts", "./src/**/*.ts", "./src/**/*.tsx", "logger.ts", "datadog.ts"]
+  "include": ["next-env.d.ts", "./src/**/*.ts", "./src/**/*.tsx", "logger.ts", "winston.ts"]
 }
